Perform revocation checks at call-site.
This is important because criticaltrust codebase only works with SignedPayload<signed_stuff>. Even though RevocationInfo also comes as SignedPayload<RevocationInfo>, this does not mean that we can simple get verified RevocationInfo and just compare that with the SignedPayload<signed_stuff>. Where signed_stuff is package manifest, release manifest etc.

Also, RevocationInfo is signed by a different key. We cannot check for revoked content unless we get the verified payload. In other words, this means that we cannot perform any comparison of bytes or strings because SignedPayload<T> will be different from T.