Clean command: removes installation from state if directory missing.
This was a bug that if the directory of the installation is
missing from the file system, and the installation is still
present in the state file, the clean command will not remove the
installation from the state file. This was certainly a problem
because if there is accidental deletion of the toolchain
directory, we need to be able to clean up the state file too.